chore: fix all clippy warnings and errors
- Fix missing `unresolved_blocker_count` field in test Task initializer
- Auto-fix 35+ collapsible_if, is_multiple_of, redundant closure, and
  other warnings via cargo clippy --fix
- Change &PathBuf to &Path in git_worktree and session_ops signatures
- Convert manual impl Future to async fn in main.rs StdioBridge
- Suppress too_many_arguments on internal functions with #[allow]

+++ src/sync/backoff.rs
@@ -36,7 +36,7 @@ impl BackoffState {
         if self.failures == 0 {
             return Duration::ZERO;
         }
-        let shifts = (self.failures - 1).min(5) as u32;
+        let shifts = (self.failures - 1).min(5);
         let secs = Self::MIN_DELAY_SECS.saturating_mul(1u64 << shifts);
         Duration::from_secs(secs.min(Self::MAX_DELAY_SECS))
     }
